+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -554,6 +554,10 @@ class GeckoEngine(
                             cookieBehavior = value.cookiePolicy.id
                         }
 
+                        if (cookieBehaviorPrivateMode != value.cookiePolicyPrivateMode.id) {
+                            cookieBehaviorPrivateMode = value.cookiePolicyPrivateMode.id
+                        }
+
                         if (cookiePurging != value.cookiePurging) {
                             setCookiePurging(value.cookiePurging)
                         }