+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/ext/TrackingProtectionPolicy.kt
@@ -19,6 +19,7 @@ fun TrackingProtectionPolicy.toContentBlockingSetting(
     enhancedTrackingProtectionLevel(getEtpLevel())
     antiTracking(getAntiTrackingPolicy())
     cookieBehavior(cookiePolicy.id)
+    cookieBehaviorPrivateMode(cookiePolicyPrivateMode.id)
     cookiePurging(cookiePurging)
     safeBrowsing(safeBrowsingPolicy.sumBy { it.id })
     strictSocialTrackingProtection(getStrictSocialTrackingProtection())