+++ mobile/android/android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/ext/TrackingProtectionPolicyKtTest.kt
@@ -26,6 +26,7 @@ class TrackingProtectionPolicyKtTest {
         assertEquals(policy.getEtpLevel(), setting.enhancedTrackingProtectionLevel)
         assertEquals(policy.getAntiTrackingPolicy(), setting.antiTrackingCategories)
         assertEquals(policy.cookiePolicy.id, setting.cookieBehavior)
+        assertEquals(policy.cookiePolicyPrivateMode.id, setting.cookieBehavior)
         assertEquals(defaultSafeBrowsing.sumBy { it.id }, setting.safeBrowsingCategories)
         assertEquals(setting.strictSocialTrackingProtection, policy.strictSocialTrackingProtection)
         assertEquals(setting.cookiePurging, policy.cookiePurging)